Replaced schlooped new parser in, all that's left is writing in all the parsing rules. I'll do that Monday.
