+++ src/compiler/mod.rs
@@ -8,7 +8,6 @@
 
 pub mod lex;
 pub mod parse;
-pub mod parse_new;
 pub mod gen;
 
 pub mod token;
